Fix for parameter hints not showing when using tabs instead of spaces
Parameter hints were not showing when using the tab character for indents.
In case of spaces a column number equals the character position in the line string, but in case of tabs that trick doesn't work.
This fix converts a column number to a character position, taking tabs into account.